pyproject.toml
[project]
name = "fastapi-auth"
version = "0.1.0"
description = ""
authors = [
    {name = "vavlxxx",email = "vladimir.nefedov357@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standart] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
package-mode=false

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
pyright = "^1.1.404"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]

[tool.ruff]
